Restructure animation modules and add transition effect
