+++ src/animation/__init__.py
@@ -1,7 +1,10 @@
-from .page import Page
-from .document import Document
-from .canvas import Canvas
-from .camera import Camera
-from .renderer import Renderer
-from .highlight import highlight_region, highlight_frames, highlight_text_frames
-from .typing_effect import typing_frames
+from .core.page import Page
+from .core.document import Document
+from .core.canvas import Canvas
+from .core.camera import Camera
+
+from .rendering.renderer import Renderer
+
+from .effects.highlight import highlight_region, highlight_frames, highlight_text_frames
+from .effects.typing_effect import typing_frames
+from .effects.transition import fade_in_frames, fade_out_frames
