+++ src/animation/core/__init__.py
@@ -1,0 +1,6 @@
+from .page import Page
+from .document import Document
+from .canvas import Canvas
+from .camera import Camera
+
+__all__ = ["Page", "Document", "Canvas", "Camera"]
