+++ src/animation/core/camera.py
