+++ src/animation/core/canvas.py
