+++ src/animation/effects/__init__.py
@@ -1,0 +1,12 @@
+from .highlight import highlight_region, highlight_frames, highlight_text_frames
+from .typing_effect import typing_frames
+from .transition import fade_in_frames, fade_out_frames
+
+__all__ = [
+    "highlight_region",
+    "highlight_frames",
+    "highlight_text_frames",
+    "typing_frames",
+    "fade_in_frames",
+    "fade_out_frames",
+]
