+++ src/animation/effects/highlight.py
