+++ src/animation/effects/typing_effect.py
