+++ src/animation/rendering/__init__.py
@@ -1,0 +1,3 @@
+from .renderer import Renderer
+
+__all__ = ["Renderer"]
